Renamed hyp3_logging -> logging in hyp3_process.daemon
